test(e2e): promote setupDeploymentMocks to shared util
Move the setupDeploymentMocks route-stub helper out of deployment-create.spec.ts into tests/utils/deployment-mocks.ts, next to the mock data it already uses, so the deployment specs can share it instead of hand-rolling the ~70 inline page.route blocks. deployment-create now imports it; behavior is unchanged.

Phase 0 (F3); the other deployment specs route through it in D1.

Refs LE-1665
